command_gen.py: replace "equiphat" with "hat"
It used to be, pre advancements, that hat was a trigger.
When people set their hat score to 1, command blocks
would set their equiphat score to 0. This was to make it easier
to type out the scoreboard name, but more importantly, to prevent
people from setting their equiphat score to anything but 0.

However, if used properly, the advancement will only be granted to people
with an equiphat score of 0.
Therefore, the scoreboard can safely be made into a trigger.
Since hat is easier to type than equiphat, rename the scoreboard.

Also, change the default equiphat value to 1.
On my server, the hat command has always been /trigger hat set 1,
and I don't want to change that.

Since 0 is unused, 0 means "invalid" now.