command_gen: add/move initialization commands to the beginning of the advancement
Add the command which adds the hat objective to the beginning.
Server owners should not have to manually set up the objective
in order to use this module.

`scoreboard enable @s hat` should go at the beginning, as the end is
reserved for players who are holding an invalid hat.