null_byte:hat/command → null_byte:hat/main
I used hat/command because it was supposed to be run by
players (with OP) or by the server.

However, since hat/main runs hat/equip for *all* players,
hat/main is unsuitable for use by a single player.

Because hat/main is the "entry point", main is thus a suitable
name.